Crate bounded_spsc_queue

Structs

Buffer
The internal memory buffer used by the queue.

Consumer
A handle to the queue which allows consuming values from the buffer

Producer
A handle to the queue which allows adding values onto the buffer

Functions

make
Creates a new SPSC Queue, returning a Producer and Consumer handle